% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcaweather.R
\name{barcaweather}
\alias{barcaweather}
\title{Obtain the temperature for any day in the coming week in Barcelona}
\usage{
barcaweather(day, metric = FALSE,
  private_key = "019b0ffe039ab1061724964decc1a73e")
}
\arguments{
\item{day}{an integer number of days from today for which temperature is desired}

\item{metric}{a boolean that determines whether the temperature returned
is in celsius (metric = TRUE) or fahrenheit (metric = FALSE)}

\item{private_key}{private key for use with darksky api}
}
\value{
a float that represents the temperature for the given day
}
\description{
The darksky api, among other functions, allows one to retrieve the temperature
of any day in the coming week. This might be useful in helping one to plan
events, or to prepare what kind of clothing to wear.
}
\details{
This package uses the darksky api to get the temperature for any day in
the coming week. The temperature is calculated by taking the mean between the
daily maximum and minimum temperatures.
}
\examples{
weather_tmr <- "1"
barcaweather(weather_tmr) # returns the weather for tomorrow in fahrenheit

}
